Raise with either call or jm
